AddDisplacement resolved

+++ functions.cpp
@@ -126,8 +126,8 @@ Add [dx,dy] random displacement to each lattice site
 void AddDisplacement(vector<vector<double>> &lattice, double delta) {
    int N = lattice.size();
    for(int j = 0; j != N-1; j++) {
-      double dx = rand()/RAND_MAX - delta/2;
-      double dy = rand()/RAND_MAX - delta/2;
+      double dx = double(rand()/RAND_MAX) - delta/2;
+      double dy = double(rand()/RAND_MAX) - delta/2;
       lattice[j] = vector<double>{dx+lattice[j][0],dy+lattice[j][1]};
    }
    return;
